+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 * Add particle config
-* Major refactor of entire project
+* Add particle editor
+* Remove individual particle toggles from settings (replaced by particle editor)
+* Improve smoke trail particles
 
 ## 1.0.2 - 2025-03-01